requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==26.1.0
backports.asyncio.runner==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.6
coverage==7.10.7
exceptiongroup==1.3.1
execnet==2.1.2
greenlet==3.2.5
idna==3.11
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
playwright==1.58.0
pluggy==1.6.0
pyee==13.0.1
Pygments==2.20.0
PyMySQL==1.1.2
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-base-url==2.1.0
pytest-cov==7.1.0
pytest-html==4.2.0
pytest-metadata==3.1.1
pytest-playwright==0.7.1
pytest-rerunfailures==16.0.1
pytest-xdist==3.8.0
python-slugify==8.0.4
requests==2.32.5
text-unidecode==1.3
tomli==2.4.1
typing_extensions==4.15.0
urllib3==2.6.3
